Begin migration to single script
I originally made all the scripts act independently but this adds tons of duplication and validation that slows things down a lot as well as making continued development difficult.
I'm going to be moving to a single (large) script which will remove duplication and speed things up.
This commit internalises a number of the external scripts for "doom_crusher.bat"