Do WAD optimisation within doom_crusher.bat
This is pretty complex due to requiring a temporary directory. We now set up one for the script as a whole so that it will be possible to run more than one instance of DOOM Crusher simultaneously (though you probably shouldn't)
The logic for WAD processing is slightly incomplete but I need to look at it with fresh eyes to handle the final details.